refactor(text-to-image): moves "Client/SDXL/conversions" to "Client/SDXL/toSharkUIOutput"
