+++ src/features/TextToImage/Client/SDXL/toSharkUIOutput/GenerateFromTextResponse.ts
