+++ src/features/TextToImage/Client/SDXL/toSharkUIOutput/StabilityAI_Client_Image.ts
