+++ src/features/TextToImage/Client/SDXL/toSharkUIOutput/TextPrompt.ts
